Re-add the casts and add a comment why they're needed

+++ src/ui_basic/button.cc
@@ -182,13 +182,14 @@ std::vector<Recti> Button::focus_overlay_rects() {
  * Redraw the button
  */
 void Button::draw(RenderTarget& dst) {
-	const bool is_flat = (enabled_ && visual_state_ == VisualState::kFlat) ||
-	                     (!enabled_ && ((disable_style_ & ButtonDisableStyle::kFlat) != 0));
+	const bool is_flat = (enabled_ && visual_state_ == VisualState::kFlat) || (!enabled_ &&
+	                     // Needs explicit cast to int to make clang-tidy happy.
+	                     (static_cast<int>(disable_style_ & ButtonDisableStyle::kFlat) != 0));
 	const bool is_permpressed =
 	   (enabled_ && visual_state_ == VisualState::kPermpressed) ||
-	   (!enabled_ && ((disable_style_ & ButtonDisableStyle::kPermpressed) != 0));
+	   (!enabled_ && (static_cast<int>(disable_style_ & ButtonDisableStyle::kPermpressed) != 0));
 	const bool is_monochrome =
-	   !enabled_ && ((disable_style_ & ButtonDisableStyle::kMonochrome) != 0);
+	   !enabled_ && (static_cast<int>(disable_style_ & ButtonDisableStyle::kMonochrome) != 0);
 
 	const UI::TextPanelStyleInfo& style_to_use =
 	   is_monochrome ? button_style().disabled() : button_style().enabled();
